requirements.txt
# Core Dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# LLM and Embeddings
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.15

# PDF Processing
pymupdf==1.23.8

# Export Functionality
reportlab==4.0.7
python-docx==1.1.0
markdown2==2.4.12

# Utilities
numpy==1.24.3
pandas==2.0.3
